+++ examples/t2-nestjs/apps/inventory/src/inventory.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Get, Post } from '@nestjs/common';
+import { Body, Controller, Get, Logger, Post } from '@nestjs/common';
 import { DataGenerator } from './utils/DataGenerator';
 import { InventoryService } from './inventory.service';
 
@@ -10,22 +10,33 @@ export class InventoryController {
 
   @Post('inventory/reservations')
   addReservation(
-    @Body() body: { sessionId: string; productId: string; units: number },
+    @Body() body: { sessionId: string; productId: string; units: number }
   ) {
-    return this.inventoryService.makeReservation(
+    Logger.log(
+      `Making reservation for ${body.productId} with ${body.units} units ...`
+    );
+    const reservation = this.inventoryService.makeReservation(
       body.productId,
       body.sessionId,
-      body.units,
+      body.units
     );
+    return {
+      id: reservation.id,
+      userId: reservation.userID,
+      productId: reservation.item.id,
+      units: reservation.units,
+    };
   }
 
   @Get('generate')
   generateData() {
     this.dataGenerator.generateProducts();
+    Logger.log('Generated products ...');
   }
 
   @Get('restock')
   restock() {
     this.dataGenerator.restockProducts();
+    Logger.log('Restocked products ...');
   }
 }