+++ examples/t2-nestjs/apps/inventory/src/inventory.module.ts
@@ -14,6 +14,7 @@ export class InventoryModule implements OnModuleInit, OnModuleDestroy {
   protected readonly client = new Kafka({
     broker: { brokers: ['localhost:9092'] },
     consumer: { groupId: 'inventory' },
+    autoRegisterTopics: true,
     registry: new Confluent({
       autoRegisterSchemas: true,
       clientConfig: {