+++ examples/t2-nestjs/apps/order/src/order.module.ts
@@ -12,6 +12,7 @@ export class OrderModule implements OnModuleInit, OnModuleDestroy {
   protected readonly client = new Kafka({
     broker: { brokers: ['localhost:9092'] },
     consumer: { groupId: 'order' },
+    autoRegisterTopics: true,
     registry: new Confluent({
       autoRegisterSchemas: true,
       clientConfig: {