+++ examples/t2-nestjs/apps/payment/src/payment.messaging.controller.ts
@@ -3,6 +3,7 @@ import { Kafka } from '@ts-messaging/client-kafka';
 import { SagaKey } from 'lib/saga/schemas/SagaKey';
 import { PaymentService } from './payment.service';
 import { PaymentSagaMessage } from './schemas/PaymentSagaMessage';
+import { SagaReply } from 'lib/saga';
 
 @Kafka.Controller()
 export class PaymentMessagingController {
@@ -16,27 +17,33 @@ export class PaymentMessagingController {
     @Kafka.Key() key: SagaKey,
     @Kafka.Value() value: PaymentSagaMessage
   ) {
-    Logger.log(
-      {
-        key,
-        value,
-      },
-      'payment.saga'
-    );
+    Logger.log({ key, value }, 'payment.saga');
 
-    await this.client.produce({
-      topic: 'order.saga.reply',
-      data: {
-        key: SagaKey,
-        value: {
-          success: this.paymentService.processPayment({
-            amount: value.total,
-            cardNumber: value.cardNumber,
-            cardOwner: value.cardOwner,
-            checksum: value.checksum,
+    try {
+      await this.client.produce({
+        topic: 'order.saga.reply',
+        data: {
+          key: key,
+          value: new SagaReply({
+            success: this.paymentService.processPayment({
+              amount: value.total,
+              cardNumber: value.cardNumber,
+              cardOwner: value.cardOwner,
+              checksum: value.checksum,
+            }),
           }),
         },
-      },
-    });
+      });
+    } catch (e) {
+      await this.client.produce({
+        topic: 'order.saga.reply',
+        data: {
+          key: key,
+          value: new SagaReply({
+            success: false,
+          }),
+        },
+      });
+    }
   }
 }