+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ts-messaging/common
 
+## 0.0.0
+
+### Patch Changes
+
+- Fixed some minor problems with schema object annotation and finished t2-nestjs example.
+
 ## 0.0.0-snapshot-20230905172825
 
 ### Patch Changes