+++ packages/registry/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ts-messaging/registry
 
+## 0.0.0
+
+### Patch Changes
+
+- Fixed some minor problems with schema object annotation and finished t2-nestjs example.
+- Updated dependencies
+  - @ts-messaging/common@0.0.0
+
 ## 0.0.0-snapshot-20230905172825
 
 ### Patch Changes