+++ packages/schema-avro/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ts-messaging/schema-avro
 
+## 0.0.0
+
+### Patch Changes
+
+- Fixed some minor problems with schema object annotation and finished t2-nestjs example.
+- Updated dependencies
+  - @ts-messaging/common@0.0.0
+  - @ts-messaging/schema@0.0.0
+
 ## 0.0.0-snapshot-20230905172825
 
 ### Patch Changes